requirements.txt
fastapi
torch
pillow
scikit-image
scikit-learn
numpy
scipy
torchvision